+++ packages/editable-html/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [9.6.1](https://github.com/pie-framework/pie-lib/compare/@pie-lib/editable-html@9.6.0...@pie-lib/editable-html@9.6.1) (2023-05-12)
+
+**Note:** Version bump only for package @pie-lib/editable-html
+
+
+
+
+
 # [9.6.0](https://github.com/pie-framework/pie-lib/compare/@pie-lib/editable-html@9.5.26...@pie-lib/editable-html@9.6.0) (2023-05-11)