+++ packages/scoring-config/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.30](https://github.com/pie-framework/pie-lib/compare/@pie-lib/scoring-config@3.6.29...@pie-lib/scoring-config@3.6.30) (2023-05-12)
+
+**Note:** Version bump only for package @pie-lib/scoring-config
+
+
+
+
+
 ## [3.6.29](https://github.com/pie-framework/pie-lib/compare/@pie-lib/scoring-config@3.6.28...@pie-lib/scoring-config@3.6.29) (2023-05-11)
 
 **Note:** Version bump only for package @pie-lib/scoring-config